requirements.txt
flask==2.3.3
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==0.8.11
markdown==3.5.1
flask-cors==4.0.0
gunicorn==21.2.0 